+++ src/biz/bokhorst/xprivacy/XActivityThread.java
@@ -180,7 +180,7 @@ public class XActivityThread extends XHook {
 					else
 						packageNames = new String[] { intent.getData().getSchemeSpecificPart() };
 					for (String packageName : packageNames)
-						if (!XPackageManager.isPackageAllowed(packageName))
+						if (!XPackageManager.isPackageAllowed(0, packageName))
 							return true;
 				}