+++ src/biz/bokhorst/xprivacy/PrivacyService.java
@@ -82,7 +82,7 @@ public class PrivacyService extends IPrivacyService.Stub {
 	private static final String cTableUsage = "usage";
 	private static final String cTableSetting = "setting";
 
-	private static final int cCurrentVersion = 453;
+	private static final int cCurrentVersion = 454;
 	private static final String cServiceName = "xprivacy453";
 
 	private boolean mCorrupt = false;