+++ CHANGELOG.md
@@ -27,6 +27,7 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 **Next release**
 
 * Broadcast *biz.bokhorst.xprivacy.action.EXCEPTION* if the database could not be read ([issue](/../../issues/2081))
+* Updated Indonesian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)