+++ CHANGELOG.md
@@ -27,6 +27,7 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 * Fixed all problems reported through the debug info
 * Updated French translation
 * Updated Indonesian translation
+* Updated Polish translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)